feat: Move risk score filter under search bar
This commit refactors the dashboard UI to improve user experience.

The key changes are:
- The "Minimum Risk Score" slider has been moved from the collapsible "Filters" panel to be permanently displayed under the main search bar in the `SearchHeader` component.
- The "Sources" filter has been removed entirely, as all data now comes from a single source (AIC).
- The `FilterPanel` component and its associated toggle button have been completely removed from the dashboard, as they are no longer needed.
- The global filter state in `zustand` has been updated to reflect these changes.
